+++ .gitignore
@@ -1,52 +1,41 @@
-.DS_STORE
-node_modules
-scripts/flow/*/.flowconfig
-*~
-*.pyc
-.grunt
-_SpecRunner.html
-__benchmarks__
+# Gradle
+.gradle/
 build/
-remote-repo/
-coverage/
-.module-cache
-fixtures/dom/public/react-dom.js
-fixtures/dom/public/react.js
-test/the-files-to-test.generated.js
-*.log*
-chrome-user-data
-*.sublime-project
-*.sublime-workspace
+
+# Eclipse
+.project
+.classpath
+.settings/
+bin/
+
+# IntelliJ
 .idea
+*.ipr
 *.iml
+*.iws
+
+# NetBeans
+nb-configuration.xml
+
+# Visual Studio Code
 .vscode
+.factorypath
+
+# OSX
+.DS_Store
+
+# Vim
 *.swp
 *.swo
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
+# patch
+*.orig
+*.rej
 
-# testing
-/coverage
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-development.env
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
+# Local environment
 .env
-/target
-.qodana
+
+# Plugin directory
+/.quarkus/cli/plugins/
+# TLS Certificates
+.certs/
