--- CHANGELOG.md
@@ -1,30 +1,0 @@
-# Changelog
-
-## [1.0.2](https://github.com/Remote-Falcon/remote-falcon-viewer/compare/v1.0.1...v1.0.2) (2024-05-04)
-
-
-### Bug Fixes
-
-* Trigger release ([20b4ee4](https://github.com/Remote-Falcon/remote-falcon-viewer/commit/20b4ee4debb3e5e3c6a78ce8ba6b0bef7a33ef60))
-
-## [1.0.1](https://github.com/Remote-Falcon/remote-falcon-viewer/compare/v1.0.0...v1.0.1) (2024-05-04)
-
-
-### Bug Fixes
-
-* Update release-please.yml ([d93b364](https://github.com/Remote-Falcon/remote-falcon-viewer/commit/d93b364789e0055912581d58b4d3bf2d772ad94a))
-* Update version.txt ([fdd8898](https://github.com/Remote-Falcon/remote-falcon-viewer/commit/fdd88985f4b7ccc2b101be3dc3066c15879a092f))
-
-## 1.0.0 (2024-05-03)
-
-
-### Bug Fixes
-
-* Update release-please.yml ([d93b364](https://github.com/Remote-Falcon/remote-falcon-viewer/commit/d93b364789e0055912581d58b4d3bf2d772ad94a))
-
-## 1.0.0 (2024-04-26)
-
-
-### Bug Fixes
-
-* Add version and changelog ([#1](https://github.com/Remote-Falcon/remote-falcon-viewer/issues/1)) ([2b8be06](https://github.com/Remote-Falcon/remote-falcon-viewer/commit/2b8be06969037451ce0475163518fde68835d105))
