+++ Dockerfile
@@ -1,15 +1,36 @@
-FROM maven:3-openjdk-17-slim AS build
-COPY src /usr/src/app/src
-COPY pom.xml /usr/src/app
-RUN mvn -f /usr/src/app/pom.xml clean package
+FROM container-registry.oracle.com/graalvm/native-image:21 AS build
+WORKDIR /app
+COPY . .
+RUN chmod +x ./gradlew
 
-FROM openjdk:17-oracle
-COPY --from=build /usr/src/app/target/remote-falcon-viewer.jar /usr/app/remote-falcon-viewer.jar
+ARG MONGO_URI
+ARG OTEL_URI
+ENV MONGO_URI=${MONGO_URI}
+ENV OTEL_URI=${OTEL_URI}
+
+RUN ./gradlew clean build -Dquarkus.native.enabled=true \
+    -Dquarkus.native.container-build=false \
+    -Dquarkus.native.builder-image=graalvm \
+    -Dquarkus.native.container-runtime=docker \
+    -Dquarkus.mongodb.connection-string=${MONGO_URI} \
+    -Dquarkus.otel.exporter.otlp.endpoint=${OTEL_URI}
+
+FROM registry.access.redhat.com/ubi9/ubi-minimal:9.2
+WORKDIR /app
+RUN chown 1001 /app && chmod "g+rwX" /app && chown 1001:root /app
+COPY --from=build --chown=1001:root /app/build/*-runner /app/application
+
+ARG MONGO_URI
+ARG OTEL_URI
+ENV MONGO_URI=${MONGO_URI}
+ENV OTEL_URI=${OTEL_URI}
+
 EXPOSE 8080
+USER 1001
 
-ARG OTEL_OPTS
-ENV OTEL_OPTS=${OTEL_OPTS}
-
-ADD 'https://dtdg.co/latest-java-tracer' /usr/app/dd-java-agent.jar
-
-ENTRYPOINT exec java $JAVA_OPTS $OTEL_OPTS -XX:FlightRecorderOptions=stackdepth=256 -XX:MaxRAMPercentage=90.0 -jar /usr/app/remote-falcon-viewer.jar
+ENTRYPOINT [ \
+"/app/application", \
+"-Dquarkus.http.host=0.0.0.0", \
+"-Dquarkus.mongodb.connection-string=${MONGO_URI}", \
+"-Dquarkus.otel.exporter.otlp.endpoint=${OTEL_URI}" \
+]
