+++ gradle.properties
@@ -1,0 +1,6 @@
+# Gradle properties
+quarkusPluginId=io.quarkus
+quarkusPluginVersion=3.17.7
+quarkusPlatformGroupId=io.quarkus.platform
+quarkusPlatformArtifactId=quarkus-bom
+quarkusPlatformVersion=3.17.7
