+++ k8s/manifest-local.yml
@@ -1,21 +1,15 @@
 ---
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: #{Namespace}#
----
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: #{Application}#
-  namespace: #{Namespace}#
+  name: remote-falcon-viewer
   labels:
-    app: #{Application}#
+    app: remote-falcon-viewer
 spec:
-  replicas: #{Replicas.Min}#
+  replicas: 4
   selector:
     matchLabels:
-      app: #{Application}#
+      app: remote-falcon-viewer
   strategy:
     type: RollingUpdate
     rollingUpdate:
@@ -24,18 +18,19 @@ spec:
   template:
     metadata:
       labels:
-        app: #{Application}#
+        app: remote-falcon-viewer
     spec:
       containers:
-        - name: #{Application}#
-          image: #{Image}#
+        - name: remote-falcon-viewer
+          imagePullPolicy: Never
+          image: remote-falcon-viewer:latest
           resources:
             requests:
-              memory: #{Requests.Memory}#
-              cpu: #{Requests.CPU}#
+              memory: 500Mi
+              cpu: 500m
             limits:
-              memory: #{Limits.Memory}#
-              cpu: #{Limits.CPU}#
+              memory: 750Mi
+              cpu: 750m
           ports:
             - containerPort: 8080
           startupProbe:
@@ -59,53 +54,45 @@ spec:
             initialDelaySeconds: 0
             failureThreshold: 30
             periodSeconds: 1
-      imagePullSecrets:
-        - name: remote-falcon-ghcr
 ---
 apiVersion: v1
 kind: Service
 metadata:
-  name: #{Application}#
-  namespace: #{Namespace}#
+  name: remote-falcon-viewer
 spec:
   type: ClusterIP
   ports:
     - port: 8080
   selector:
-    app: #{Application}#
+    app: remote-falcon-viewer
 ---
 apiVersion: networking.k8s.io/v1
 kind: Ingress
 metadata:
-  name: #{Application}#
-  namespace: #{Namespace}#
-  annotations:
-    kubernetes.io/ingress.class: nginx
-    nginx.ingress.kubernetes.io/use-forwarded-headers: "true"
+  name: remote-falcon-viewer
 spec:
   rules:
-    - host: #{Host}#
-      http:
+    - http:
         paths:
-          - pathType: Prefix
-            path: /remote-falcon-viewer
+          - path: /remote-falcon-viewer
+            pathType: Prefix
             backend:
               service:
-                name: #{Application}#
+                name: remote-falcon-viewer
                 port:
                   number: 8080
 ---
 apiVersion: autoscaling/v2
 kind: HorizontalPodAutoscaler
 metadata:
-  name: #{Application}#
+  name: remote-falcon-viewer
 spec:
   scaleTargetRef:
     apiVersion: apps/v1
     kind: Deployment
-    name: #{Application}#
-  minReplicas: #{Replicas.Min}#
-  maxReplicas: #{Replicas.Max}#
+    name: remote-falcon-viewer
+  minReplicas: 4
+  maxReplicas: 6
   metrics:
     - type: Resource
       resource:
