+++ src/main/java/com/remotefalcon/util/LocationUtil.java
@@ -1,0 +1,15 @@
+package com.remotefalcon.util;
+
+public class LocationUtil {
+    public static Double asTheCrowFlies(double lat1, double lon1, double lat2, double lon2) {
+        if ((lat1 == lat2) && (lon1 == lon2)) {
+            return 0.0;
+        }else {
+            double theta = lon1 - lon2;
+            double dist = Math.sin(Math.toRadians(lat1)) * Math.sin(Math.toRadians(lat2)) + Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) * Math.cos(Math.toRadians(theta));
+            dist = Math.acos(dist);
+            dist = Math.toDegrees(dist);
+            return dist * 60 * 1.1515;
+        }
+    }
+}
