--- src/main/java/com/remotefalcon/viewer/Application.java
@@ -1,23 +1,0 @@
-package com.remotefalcon.viewer;
-
-import org.dozer.DozerBeanMapper;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
-@EnableAspectJAutoProxy
-@EnableMongoRepositories
-public class Application {
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-
-  @Bean
-  public DozerBeanMapper mapper() {
-    return new DozerBeanMapper();
-  }
-}
