--- src/main/java/com/remotefalcon/viewer/aop/AccessAspect.java
@@ -1,37 +1,0 @@
-package com.remotefalcon.viewer.aop;
-
-import com.remotefalcon.library.enums.StatusResponse;
-import com.remotefalcon.viewer.util.AuthUtil;
-import jakarta.servlet.http.HttpServletRequest;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-@Aspect
-@Component
-public class AccessAspect {
-  @Autowired
-  private AuthUtil authUtil;
-
-  @Around("@annotation(RequiresAccess)")
-  public Object isJwtValid(ProceedingJoinPoint proceedingJoinPoint) throws Throwable {
-    HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest();
-    if(this.authUtil.isJwtValid(request)) {
-      return proceedingJoinPoint.proceed();
-    }
-    throw new RuntimeException(StatusResponse.INVALID_JWT.name());
-  }
-
-  @Around("@annotation(RequiresAPIAccess)")
-  public Object isApiJwtValid(ProceedingJoinPoint proceedingJoinPoint) throws Throwable {
-    HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest();
-    if(this.authUtil.isApiJwtValid(request)) {
-      return proceedingJoinPoint.proceed();
-    }
-    throw new RuntimeException(StatusResponse.INVALID_JWT.name());
-  }
-}
