--- src/main/java/com/remotefalcon/viewer/aop/RequiresAccess.java
@@ -1,11 +1,0 @@
-package com.remotefalcon.viewer.aop;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequiresAccess {
-}
