--- src/main/java/com/remotefalcon/viewer/configuration/AsyncConfiguration.java
@@ -1,18 +1,0 @@
-package com.remotefalcon.viewer.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-@EnableAsync
-public class AsyncConfiguration  {
-  @Bean
-  public Executor asyncExecutor() {
-    ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-    return executor;
-  }
-}
