--- src/main/java/com/remotefalcon/viewer/configuration/CustomExceptionResolver.java
@@ -1,20 +1,0 @@
-package com.remotefalcon.viewer.configuration;
-
-import graphql.GraphQLError;
-import graphql.GraphqlErrorBuilder;
-import graphql.schema.DataFetchingEnvironment;
-import org.springframework.graphql.execution.DataFetcherExceptionResolverAdapter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class CustomExceptionResolver extends DataFetcherExceptionResolverAdapter {
-
-    @Override
-    protected GraphQLError resolveToSingleError(Throwable ex, DataFetchingEnvironment env) {
-        return GraphqlErrorBuilder.newError()
-                .message(ex.getMessage())
-                .path(env.getExecutionStepInfo().getPath())
-                .location(env.getField().getSourceLocation())
-                .build();
-    }
-}
