--- src/main/java/com/remotefalcon/viewer/configuration/GraphQLConfig.java
@@ -1,16 +1,0 @@
-package com.remotefalcon.viewer.configuration;
-
-import graphql.scalars.ExtendedScalars;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.graphql.execution.RuntimeWiringConfigurer;
-
-@Configuration
-public class GraphQLConfig {
-    @Bean
-    public RuntimeWiringConfigurer runtimeWiringConfigurer() {
-        return wiringBuilder -> wiringBuilder
-                .scalar(ExtendedScalars.GraphQLLong);
-
-    }
-}
