--- src/main/java/com/remotefalcon/viewer/dto/RequestVoteResponse.java
@@ -1,14 +1,0 @@
-package com.remotefalcon.viewer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RequestVoteResponse {
-    private String message;
-}
