--- src/main/java/com/remotefalcon/viewer/dto/TokenDTO.java
@@ -1,14 +1,0 @@
-package com.remotefalcon.viewer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenDTO {
-  private String showSubdomain;
-}
