--- src/main/java/com/remotefalcon/viewer/util/ClientUtil.java
@@ -1,24 +1,0 @@
-package com.remotefalcon.viewer.util;
-
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-@Service
-@Slf4j
-public class ClientUtil {
-  @Value("${client.header}")
-  String clientHeader;
-
-  public String getClientIp(HttpServletRequest request) {
-    String remoteAddr = "";
-    if (request != null) {
-      remoteAddr = request.getHeader(clientHeader);
-      if (remoteAddr == null || "".equals(remoteAddr)) {
-        remoteAddr = request.getRemoteAddr();
-      }
-    }
-    return remoteAddr;
-  }
-}
