--- src/main/java/com/remotefalcon/viewer/util/LogConstants.java
@@ -1,8 +1,0 @@
-package com.remotefalcon.viewer.util;
-
-public class LogConstants {
-    public static final String SEQUENCE_REQUESTED_SUCCESS = "Sequence Requested Successfully";
-    public static final String SEQUENCE_GROUP_REQUESTED_SUCCESS = "Sequence Group Requested Successfully";
-    public static final String SEQUENCE_VOTED_SUCCESS = "Sequence Voted Successfully";
-    public static final String SEQUENCE_GROUP_VOTED_SUCCESS = "Sequence Group Voted Successfully";
-}
