--- src/main/resources/application-local.yml
@@ -1,5 +1,0 @@
-server:
-  port: 8082
-
-client:
-  header: "CF-Connecting-IP"
