+++ src/main/resources/application.properties
@@ -1,0 +1,18 @@
+quarkus.package.jar.enabled=false
+quarkus.http.root-path=/remote-falcon-viewer
+quarkus.http.port=8080
+quarkus.http.proxy.allow-forwarded=true
+quarkus.http.proxy.proxy-address-forwarding=true
+quarkus.http.proxy.enable-forwarded-host=true
+quarkus.http.proxy.enable-forwarded-prefix=true
+
+#MongoDB
+quarkus.mongodb.database=remote-falcon
+
+#Component Scans
+quarkus.index-dependency.remote-falcon-library.group-id=com.github.Remote-Falcon
+quarkus.index-dependency.remote-falcon-library.artifact-id=remote-falcon-library
+
+#OpTel
+quarkus.otel.metrics.enabled=true
+quarkus.application.name=remote-falcon-viewer
