--- src/main/resources/graphql/inputs.graphqls
@@ -1,48 +1,0 @@
-input SequenceInput {
-    name: String
-    key: Long
-    displayName: String
-    duration: Int
-    visible: Boolean
-    index: Int
-    order: Int
-    imageUrl: String
-    active: Boolean
-    visibilityCount: Int
-    type: String
-    group: String
-    category: String
-    artist: String
-    total: Int
-}
-
-input SequenceGroupInput {
-    name: String
-    visibilityCount: Int
-}
-
-input RequestInput {
-    sequence: SequenceInput
-    position: Int
-    ownerRequested: Boolean
-}
-
-input VoteInput {
-    sequence: SequenceInput
-    votes: Int
-    viewersVoted: [String]
-    lastVoteTime: String
-    ownerVoted: Boolean
-}
-
-input PsaSequenceInput {
-    name: String
-    order: Int
-    lastPlayed: String
-}
-
-input PageInput {
-    name: String
-    active: Boolean
-    html: String
-}
