--- src/main/resources/graphql/query.graphqls
@@ -1,4 +1,0 @@
-type Query {
-    getShow: Show
-    activeViewerPage: String
-}
