--- src/main/resources/graphql/scalar.graphqls
@@ -1,1 +1,0 @@
-scalar Long
