--- src/main/resources/graphql/types.graphqls
@@ -1,98 +1,0 @@
-type Show {
-    showSubdomain: String
-    showName: String
-
-    preferences: Preference
-    sequences: [Sequence]
-    sequenceGroups: [SequenceGroup]
-    pages: [Page]
-    requests: [Request]
-    votes: [Vote]
-    activeViewers: [ActiveViewer]
-    playingNow: String
-    playingNext: String
-    playingNowSequence: Sequence
-    playingNextSequence: Sequence
-    playingNextFromSchedule: String
-}
-
-type Preference {
-    viewerControlEnabled: Boolean
-    viewerControlMode: String
-    resetVotes: Boolean
-    jukeboxDepth: Int
-    locationCheckMethod: String
-    showLatitude: Float
-    showLongitude: Float
-    allowedRadius: Float
-    checkIfVoted: Boolean
-    checkIfRequested: Boolean
-    psaEnabled: Boolean
-    psaFrequency: Int
-    jukeboxRequestLimit: Int
-    locationCode: Int
-    hideSequenceCount: Int
-    makeItSnow: Boolean
-    managePsa: Boolean
-    sequencesPlayed: Int
-    pageTitle: String
-    pageIconUrl: String
-    selfHostedRedirectUrl: String
-}
-
-type Sequence {
-    name: String
-    key: Long
-    displayName: String
-    duration: Int
-    visible: Boolean
-    index: Int
-    order: Int
-    imageUrl: String
-    active: Boolean
-    visibilityCount: Int
-    type: String
-    group: String
-    category: String
-    artist: String
-}
-
-type SequenceGroup {
-    name: String
-    votes: Int
-    lastVoteTime: String
-    totalVotes: Int
-    visibilityCount: Int
-}
-
-type Request {
-    sequence: Sequence
-    position: Int
-    ownerRequested: Boolean
-}
-
-type Vote {
-    sequence: Sequence
-    sequenceGroup: SequenceGroup
-    votes: Int
-    viewersVoted: [String]
-    lastVoteTime: String
-    ownerVoted: Boolean
-}
-
-type ActiveViewer {
-    ipAddress: String
-    visitDateTime: String
-}
-
-type PsaSequence {
-    name: String
-    order: Int
-    lastPlayed: String
-}
-
-type Page {
-    name: String
-    active: Boolean
-    html: String
-}
