+++ src/native-test/java/com/remotefalcon/ExampleResourceIT.java
@@ -1,0 +1,8 @@
+package com.remotefalcon;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+class ExampleResourceIT extends ExampleResourceTest {
+    // Execute the same tests but in packaged mode.
+}
