+++ src/test/java/com/remotefalcon/ExampleResourceTest.java
@@ -1,0 +1,20 @@
+package com.remotefalcon;
+
+import io.quarkus.test.junit.QuarkusTest;
+import org.junit.jupiter.api.Test;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.CoreMatchers.is;
+
+@QuarkusTest
+class ExampleResourceTest {
+//    @Test
+//    void testHelloEndpoint() {
+//        given()
+//                .when().get("/hello")
+//                .then()
+//                .statusCode(200)
+//                .body(is("Hello RESTEasy"));
+//    }
+
+}
